+++ cli/src/merge_tools/diff_working_copies.rs
@@ -10,6 +10,7 @@ use futures::StreamExt as _;
 use jj_lib::backend::MergedTreeId;
 use jj_lib::conflicts::ConflictMarkerStyle;
 use jj_lib::fsmonitor::FsmonitorSettings;
+use jj_lib::gitattributes::GitAttributesFile;
 use jj_lib::gitignore::GitIgnoreFile;
 use jj_lib::local_working_copy::EolConversionMode;
 use jj_lib::local_working_copy::TreeState;
@@ -270,6 +271,7 @@ diff editing in mind and be a little inaccurate.
     pub fn snapshot_results(
         self,
         base_ignores: Arc<GitIgnoreFile>,
+        base_attributes: Arc<GitAttributesFile>,
     ) -> Result<MergedTreeId, DiffEditError> {
         if let Some(path) = self.instructions_path_to_cleanup {
             std::fs::remove_file(path).ok();
@@ -280,6 +282,7 @@ diff editing in mind and be a little inaccurate.
         let mut output_tree_state = diff_wc.output.unwrap_or(diff_wc.right);
         output_tree_state.snapshot(&SnapshotOptions {
             base_ignores,
+            base_attributes,
             progress: None,
             start_tracking_matcher: &EverythingMatcher,
             max_new_file_size: u64::MAX,