+++ cli/src/merge_tools/mod.rs
@@ -29,6 +29,7 @@ use jj_lib::config::ConfigNamePathBuf;
 use jj_lib::conflicts::ConflictMarkerStyle;
 use jj_lib::conflicts::MaterializedFileConflictValue;
 use jj_lib::conflicts::try_materialize_file_conflict_value;
+use jj_lib::gitattributes::GitAttributesFile;
 use jj_lib::gitignore::GitIgnoreFile;
 use jj_lib::matchers::Matcher;
 use jj_lib::merge::Merge;
@@ -235,6 +236,7 @@ pub fn get_external_tool_config(
 pub struct DiffEditor {
     tool: DiffEditTool,
     base_ignores: Arc<GitIgnoreFile>,
+    base_attributes: Arc<GitAttributesFile>,
     use_instructions: bool,
     conflict_marker_style: ConflictMarkerStyle,
 }
@@ -246,11 +248,19 @@ impl DiffEditor {
         name: &str,
         settings: &UserSettings,
         base_ignores: Arc<GitIgnoreFile>,
+        base_attributes: Arc<GitAttributesFile>,
         conflict_marker_style: ConflictMarkerStyle,
     ) -> Result<Self, MergeToolConfigError> {
         let tool = DiffEditTool::get_tool_config(settings, name)?
             .unwrap_or_else(|| DiffEditTool::external(ExternalMergeTool::with_program(name)));
-        Self::new_inner(name, tool, settings, base_ignores, conflict_marker_style)
+        Self::new_inner(
+            name,
+            tool,
+            settings,
+            base_ignores,
+            base_attributes,
+            conflict_marker_style,
+        )
     }
 
     /// Loads the default diff editor from the settings.
@@ -258,6 +268,7 @@ impl DiffEditor {
         ui: &Ui,
         settings: &UserSettings,
         base_ignores: Arc<GitIgnoreFile>,
+        base_attributes: Arc<GitAttributesFile>,
         conflict_marker_style: ConflictMarkerStyle,
     ) -> Result<Self, MergeToolConfigError> {
         let args = editor_args_from_settings(ui, settings, "ui.diff-editor")?;
@@ -267,7 +278,14 @@ impl DiffEditor {
             None
         }
         .unwrap_or_else(|| DiffEditTool::external(ExternalMergeTool::with_edit_args(&args)));
-        Self::new_inner(&args, tool, settings, base_ignores, conflict_marker_style)
+        Self::new_inner(
+            &args,
+            tool,
+            settings,
+            base_ignores,
+            base_attributes,
+            conflict_marker_style,
+        )
     }
 
     fn new_inner(
@@ -275,6 +293,7 @@ impl DiffEditor {
         tool: DiffEditTool,
         settings: &UserSettings,
         base_ignores: Arc<GitIgnoreFile>,
+        base_attributes: Arc<GitAttributesFile>,
         conflict_marker_style: ConflictMarkerStyle,
     ) -> Result<Self, MergeToolConfigError> {
         if let DiffEditTool::External(mergetool) = &tool
@@ -287,6 +306,7 @@ impl DiffEditor {
         Ok(Self {
             tool,
             base_ignores,
+            base_attributes,
             use_instructions: settings.get_bool("ui.diff-instructions")?,
             conflict_marker_style,
         })
@@ -314,6 +334,7 @@ impl DiffEditor {
                     matcher,
                     instructions.as_deref(),
                     self.base_ignores.clone(),
+                    self.base_attributes.clone(),
                     self.conflict_marker_style,
                 )
             }
@@ -508,6 +529,7 @@ mod tests {
                 name,
                 &settings,
                 GitIgnoreFile::empty(),
+                Arc::new(GitAttributesFile::default()),
                 ConflictMarkerStyle::Diff,
             )
             .map(|editor| editor.tool)
@@ -585,6 +607,7 @@ mod tests {
                 &ui,
                 &settings,
                 GitIgnoreFile::empty(),
+                Arc::new(GitAttributesFile::default()),
                 ConflictMarkerStyle::Diff,
             )
             .map(|editor| editor.tool)