+++ CHANGELOG.md
@@ -626,6 +626,9 @@ Thanks to the people who made this release happen!
 * Evolution history is now stored in the operation log. `jj evolog` can show
   associated operations for commits created by new jj versions.
 
+* Added `git.ignore-filters` setting to specify what filtered files in
+  `.gitattributes` are ignored by `jj`. Defaults to `["lfs"]`.
+
 ### Fixed bugs
 
 * Work around a git issue that could cause subprocess operations to hang if the